flask/requirements.txt
Pillow
flask
tensorflow==2.0.0
protobuf==3.6.0